Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/full-block-iframe-no-inherit-expected.txt

CONSOLE MESSAGE: line 4: The XSS Auditor refused to execute a script in 'http://127.0.0.1:8000/security/xssAuditor/resources/echo-intertag.pl?test=/security/xssAuditor/full-block-iframe-no-inherit.html&q=%3Cscript%3Ealert(/XSS/)%3C/script%3E%3Cp%3EIf%20you%20see%20this%20message%20and%20no%20JavaScript%20alert()%20then%20the%20test%20PASSED.%3C/p%3E' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
This tests that the header X-XSS-Protection is not inherited by the iframe below:



--------
Frame: 'frame'
--------
If you see this message and no JavaScript alert() then the test PASSED.
